requirements.txt
beautifulsoup4
Django
django-jazzmin
Pillow
psycopg2-binary
python-dotenv
requests